shuffle before funcs into middleware package
